tests for createCOllectUrl and refactor
